moved queries folder under grapgh-module folder
