+++ console/console-init/ui/src/graphql-module/queries/AddressSpace.ts
