+++ console/console-init/ui/src/graphql-module/queries/Connection.ts
