+++ console/console-init/ui/src/graphql-module/queries/NameSpace.ts
