+++ console/console-init/ui/src/graphql-module/queries/index.ts
