+++ console/console-init/ui/src/pages/AddressSpaceDetail/ConnectionList/ConnectionListFilter.tsx
@@ -31,7 +31,7 @@ import useWindowDimensions from "components/common/WindowDimension";
 import { ISortBy } from "@patternfly/react-table";
 import { SortForMobileView } from "components/common/SortForMobileView";
 import { useApolloClient } from "@apollo/react-hooks";
-import { RETURN_ALL_CONNECTIONS_HOSTNAME_AND_CONTAINERID_OF_ADDRESS_SPACES_FOR_TYPEAHEAD_SEARCH } from "queries";
+import { RETURN_ALL_CONNECTIONS_HOSTNAME_AND_CONTAINERID_OF_ADDRESS_SPACES_FOR_TYPEAHEAD_SEARCH } from "graphql-module/queries";
 import { IConnectionListNameSearchResponse } from "types/ResponseTypes";
 import {
   TypeAheadMessage,
