+++ console/console-init/ui/src/pages/ConnectionDetail/ConnectionLinksFilter.tsx
@@ -38,7 +38,7 @@ import {
 import {
   RETURN_ALL_CONNECTION_LINKS_FOR_NAME_SEARCH,
   RETURN_ALL_CONNECTION_LINKS_FOR_ADDRESS_SEARCH
-} from "queries";
+} from "graphql-module/queries";
 import {
   TypeAheadMessage,
   MAX_ITEM_TO_DISPLAY_IN_TYPEAHEAD_DROPDOWN,
