+++ console/console-init/ui/src/pages/CreateAddressSpace/CreateAddressSpacePage.tsx
@@ -7,7 +7,7 @@ import * as React from "react";
 import { Wizard } from "@patternfly/react-core";
 import { AddressSpaceConfiguration } from "pages/CreateAddressSpace/CreateAddressSpaceConfiguration";
 import { ReviewAddressSpace } from "pages/CreateAddressSpace/ReviewAddressSpace";
-import { CREATE_ADDRESS_SPACE } from "queries";
+import { CREATE_ADDRESS_SPACE } from "graphql-module/queries";
 import { useMutationQuery } from "hooks";
 
 interface ICreateAddressSpaceProps {
